refactor(pitchfork): refactor lib
refactor the rusty_pitchfork library to have a core manager,
DomoPitchfork, and move individual API impls into their
respective domo mods.

BREAKING CHANGE:
crate name changed to domo_pitchfork.
RustyPitchfork changed to DomoPitchfork.
DomoPitchfork has fn to generate API Request Builders.
The API Request Builders have all the same previous functionallity,
but method names have changed.

+++ rusty-pitchfork/src/domo/policy.rs
@@ -2,14 +2,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Policy {
-    pub id: i32,
+    pub id: u32,
     pub name: String,
     #[serde(rename = "type")]
     pub policy_type: String,
     #[serde(rename = "users")]
-    pub user_ids: Vec<i32>,
+    pub user_ids: Vec<u64>,
     #[serde(rename = "groups")]
-    pub group_ids: Vec<i32>,
+    pub group_ids: Vec<u64>,
     pub filters: Vec<Filter>,
 }
